pyproject.toml
[project]
name = "agenticrag-mcp-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx>=0.2.4",
    "faiss-cpu>=1.11.0",
    "groq>=0.29.0",
    "numpy>=2.3.1",
    "openai>=1.93.2",
    "pandas>=2.3.1",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
]
